style: keep the exit helpers out of Process.kt
Moves killTree, destroyAndAwaitExit and its polling helper to their own file:
Process.kt was over the per-file function threshold once both killTree and
readStderrSnapshot were added. Also folds one early return and wraps a long
line.

No behaviour change. Lint issues on these files are back to the same count as
main.
