+++ core/src/commonMain/kotlin/dev/kdriver/core/browser/DefaultBrowser.kt
@@ -366,7 +366,9 @@ open class DefaultBrowser(
         // reach the browser's child processes, which keep open handles on the profile directory —
         // a browser started on the same --user-data-dir right after would hang before opening its
         // debug port, and the start would time out for no visible reason.
-        process?.let { if (!it.destroyAndAwaitExit()) logger.warn("Browser process ${it.pid()} still alive after kill") }
+        process?.let {
+            if (!it.destroyAndAwaitExit()) logger.warn("Browser process ${it.pid()} still alive after kill")
+        }
         process = null
         logger.debug("Closing connection...")
         connection?.close()
